+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 Github 源码：https://github.com/giscafer/vscode-frontend-box
 
+## [v0.1.7]
+
+- 添加了微信阅读，`CTRL+P` 输入 `weread` 即可打开；
+- 新增配置项 `frontend-box.disableToolBar` ，用于设定需要禁用Browser Preview toolBar 的网址列表；
+- 修复了博客列表解析md异常的问题，使用 `raw.fasthub.org` 代理 github raw 地址，便于国内访问；
+- 同步上游Browser Preivew最新源码的build包;
+
 ## [v0.1.6]
 
 - github 资源更换为 `https://raw.githubusercontents.com/` 域名